+++ app/Http/Controllers/Admin/VehicleController.php
@@ -88,7 +88,7 @@ class VehicleController extends Controller
      */
     public function show(Vehicle $vehicle)
     {
-        $vehicle->load(['vehicleType', 'brand', 'carModel', 'equipment.equipmentType', 'issues', 'deadlines']);
+        $vehicle->load(['vehicleType', 'brand', 'carModel', 'equipment.equipmentType', 'issues', 'deadlines', 'mileageLogs']);
 
         $vehicleAppointments = $vehicle->maintenanceRecords()
             ->with('issue', 'provider')